refactored user/exercise events,api,ui
all user actions work. exercise actions do not work yet, however, have
the same formatting as the user actions. Might be a problem requiring
exercise actions modules. There was a lot of renaming of variabels
and files.
