--- assets/scripts/exercise/events.js
@@ -1,3 +1,0 @@
-'use strict'
-const addExercise = require('./exerciseActions/addExercise')
-  $('#addExerciseButton').on('submit', addExercise.check)
