+++ assets/scripts/index.js
@@ -2,8 +2,9 @@
 
 const setAPIOrigin = require('../../lib/set-api-origin')
 const config = require('./config')
-const showHide = require('./../showHide')
-const userEvents = require('./user/events')
+const userShowHide = require('./user/userShowHide')
+const userEvents = require('./user/userEvents')
+const exerciseEvents = require('./exercise/exerciseEvents')
 
 $(() => {
   setAPIOrigin(location, config)
@@ -12,11 +13,13 @@ $(() => {
 // use require with a reference to bundle the file and use it in this file
 // const example = require('./example')
 $(() => {
-  showHide.frontPage()
-  userEvents.addHandlers()
+  userShowHide.frontPage()
+  userEvents.addUserHandlers()
+  exerciseEvents.exerciseHandlers()
 })
 // use require without a reference to ensure a file is bundled
 require('./example')
+
 $(() => {
 
 })
