+++ assets/scripts/user/userShowHide.js
