+++ index.html
@@ -162,15 +162,11 @@
       <div class="modal fade" id="addExerciseModal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel">
         <div class="modal-dialog" role="document">
           <div class="modal-content">
-
             <div class="modal-header">
               <button type="button" class="close buttonCloseSignUp" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-              <h4 class="modal-title" id="myModalLabel-signUp">Add Exercises
-    </h4>
+              <h4 class="modal-title" id="myModalLabel-signUp">Add Exercises</h4>
               <h5 id="addExerciseFailure"> </h5>
-
             </div>
-
             <div class="modal-body">
               <form id="addExerciseForm" class="form-horizontal">
                 <div class="form-group">
@@ -202,7 +198,7 @@
 <button id="resetButton" class="btn btn-primary btn-lg">
 Delete Exercise
 </button>
-<button type="button" id="showAllExercises" class="btn btn-primary btn-lg">
+<button type="button" id="showAllExercisesButton" class="btn btn-primary btn-lg">
 Show All Exercises
 </button>
 </nav>
